Bring in ExtendedDcMotor, change APIs for fractional encoder counts

+++ TeamCode/src/main/java/com/hfrobots/tnt/corelib/drive/DualDcMotor.java
@@ -230,7 +230,7 @@ public class DualDcMotor implements ExtendedDcMotor {
     }
 
     @Override
-    public int getEncoderCountsPerRevolution() {
+    public double getEncoderCountsPerRevolution() {
         return firstMotor.getEncoderCountsPerRevolution();
     }
